requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
audioop-lts==0.2.1
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.0
cairocffi==1.7.1
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cssselect2==0.7.0
distro==1.9.0
Flask==3.1.0
fonttools==4.55.3
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-core==2.4.1
google-cloud-speech==2.30.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.22.0
nltk==3.9.1
numpy==2.2.2
openai==1.59.8
packaging==24.2
pandas==2.2.3
pango==0.0.1
pillow==11.1.0
proto-plus==1.25.0
protobuf==5.29.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
pydyf==0.11.0
Pygments==2.19.1
pyphen==0.17.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.36.1
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
scipy==1.15.1
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.1
SpeechRecognition==3.14.0
standard-aifc==3.13.0
standard-chunk==3.13.0
streamlit==1.41.1
tenacity==9.0.0
textblob==0.19.0
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
watchdog==6.0.0
weasyprint==63.1
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
zopfli==0.2.3.post1
